backend/foodgram/api/mixins.py
from rest_framework import mixins, viewsets


class ReadListOrObjectViewSet(
    mixins.ListModelMixin,
    viewsets.GenericViewSet,
    mixins.RetrieveModelMixin
):
    pass


class ReadOrCreateViewSet(
    mixins.CreateModelMixin,
    mixins.ListModelMixin,
    viewsets.GenericViewSet
):
    pass
