backend/foodgram/requirements.txt
asgiref==3.5.2
Django==2.2.19
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
psycopg2-binary==2.8.6 
python-dotenv
gunicorn